Fixup: move error file into app
