+++ internal/app/errors.go
@@ -1,8 +1,7 @@
-package store
+package app
 
 import "errors"
 
 var (
-	// RecordNotFound ...
 	RecordNotFound = errors.New("record not found")
 )
